+++ R/oa_fetch.R
@@ -14,7 +14,7 @@ oa_entities <- function() {
 #' @inheritParams oa_query
 #' @inheritParams oa_request
 #' @param abstract Logical. If TRUE, the function returns also the abstract of each item.
-#' Default to \code{abstract = FALSE}.
+#' Default to \code{abstract = TRUE}.
 #' The argument is ignored if entity is different from "works".
 #' @param output Character. Type of output, either a list or a tibble/data.frame.
 #'
@@ -67,6 +67,8 @@ oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(shorten
                      abstract = TRUE,
                      endpoint = "https://api.openalex.org",
                      per_page = 200,
+                     paging = NULL,
+                     pages = NULL,
                      count_only = FALSE,
                      mailto = oa_email(),
                      api_key = oa_apikey(),
@@ -77,9 +79,9 @@ oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(shorten
   if (output == "dataframe") output <- "tibble"
   filter <- list(...)
 
-  # if multiple identifiers are provided, use openalex_id or doi as a filter attribute
+  # if multiple identifiers are provided, use openalex or doi as a filter attribute
   multiple_id <- length(identifier) > 1
-  if (multiple_id) filter <- c(filter, list(openalex_id = identifier))
+  if (multiple_id) filter <- c(filter, list(openalex = identifier))
 
   # overcome OA limitation of combining 50 values (OR) for a filter at a time
   # https://docs.openalex.org/how-to-use-the-api/get-lists-of-entities/filter-entity-lists#addition-or
@@ -96,7 +98,9 @@ oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(shorten
 
   if (!is.null(options$sample) && (options$sample > per_page)) {
     paging <- "page"
-  } else {
+  } else if (!is.null(options$page)){
+    paging <- "page"
+  } else if (is.null(paging)){
     paging <- "cursor"
   }
 
@@ -122,6 +126,7 @@ oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(shorten
       ),
       per_page = per_page,
       paging = paging,
+      pages = pages,
       count_only = count_only,
       mailto = mailto,
       api_key = api_key,
@@ -130,7 +135,6 @@ oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(shorten
   }
 
   if (length(final_res[[1]]) == 0) { # || is.null(final_res[[1]][[1]]$id)
-    warning("No collection found!")
     return(NULL)
   }
 
@@ -139,7 +143,7 @@ oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(shorten
   } else {
     do.call(rbind, lapply(
       final_res, oa2df,
-      entity = entity, abstract = abstract,
+      entity = entity, options = options, abstract = abstract,
       count_only = count_only, group_by = group_by,
       verbose = verbose
     ))
@@ -161,6 +165,12 @@ oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(shorten
 #' Defaults to 200.
 #' @param paging Character.
 #' Either "cursor" for cursor paging or "page" for basic paging.
+#' When used with `options$sample` and or `pages`,
+#' paging is also automatically set to basic paging: `paging = "page"`
+#' to avoid duplicates and get the right page.
+#' See https://docs.openalex.org/how-to-use-the-api/get-lists-of-entities/paging.
+#' @param pages Integer vector.
+#' The range of pages to return. If NULL, return all pages.
 #' @param count_only Logical.
 #' If TRUE, the function returns only the number of item matching the query.
 #' Defaults to FALSE.
@@ -301,6 +311,7 @@ oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(shorten
 oa_request <- function(query_url,
                        per_page = 200,
                        paging = "cursor",
+                       pages = NULL,
                        count_only = FALSE,
                        mailto = oa_email(),
                        api_key = oa_apikey(),
@@ -335,13 +346,22 @@ oa_request <- function(query_url,
   } else {
     return(res)
   }
+  n_items <- res$meta$count
+  n_pages <- ceiling(n_items / per_page)
 
   ## number of pages
-  n_items <- res$meta$count
-  n_pages <- ceiling(res$meta$count / per_page)
-  pages <- seq.int(n_pages)
+  if (is.null(pages)){
+    pages <- seq.int(n_pages)
+  } else {
+    pages <- pages[pages <= n_pages]
+    n_pages <- length(pages)
+    n_items <- min(n_items - per_page * (utils::tail(pages, 1) - n_pages), per_page * n_pages)
+    message("Using basic paging...")
+    paging <- "page"
+  }
 
-  if (n_items <= 0) {
+  if (n_items <= 0 || n_pages <= 0) {
+    warning("No records found!")
     return(list())
   }
 
@@ -360,14 +380,14 @@ oa_request <- function(query_url,
   query_ls[["per-page"]] <- per_page
 
   # Activation of cursor pagination
-  next_page <- get_next_page(paging, 1)
   data <- vector("list", length = n_pages)
+  res <- NULL
   for (i in pages) {
     if (verbose) pb$tick()
     Sys.sleep(1 / 100)
+    next_page <- get_next_page(paging, i, res)
     query_ls[[paging]] <- next_page
     res <- api_request(query_url, ua, query = query_ls)
-    next_page <- get_next_page(paging, i + 1, res)
     if (!is.null(res$results)) data[[i]] <- res$results
   }
 
@@ -405,21 +425,26 @@ get_next_page <- function(paging, i, res = NULL) {
 #' The argument can be one of c("works", "authors", "venues", "institutions", "concepts").
 #' If not provided, `entity` is guessed from `identifier`.
 #' @param options List. Additional parameters to add in the query. For example:
+#'
 #' - `select` Character vector. Top-level fields to show in output.
 #' Defaults to NULL, which returns all fields.
 #' https://docs.openalex.org/how-to-use-the-api/get-single-entities/select-fields
+#'
 #' - `sort` Character. Attribute to sort by.
 #' For example: "display_name" for venues or "cited_by_count:desc" for works.
 #' See more at <https://docs.openalex.org/how-to-use-the-api/get-lists-of-entities/sort-entity-lists>.
+#'
 #' - `sample` Integer. Number of (random) records to return.
 #' Should be no larger than 10,000.
 #' Defaults to NULL, which returns all records satisfying the query.
 #' Read more at <https://docs.openalex.org/how-to-use-the-api/get-lists-of-entities/sample-entity-lists>.
+#'
 #' - `seed` Integer.
 #' A seed value in order to retrieve the same set of random records in
 #' the same order when used multiple times with `sample`.
 #' IMPORTANT NOTE: Depending on your query, random results with a seed value may change over time due to new records coming into OpenAlex.
 #' This argument is likely only useful when queries happen close together (within a day).
+#'
 #' @param search Character. Search is just another kind of filter, one that all five endpoints support.
 #' But unlike the other filters, search does NOT require an exact match.
 #' This is particularly useful in author queries because many authors have middle names, which may not exist or do so in a variety of forms.
@@ -518,9 +543,6 @@ oa_query <- function(filter = NULL,
                      endpoint = "https://api.openalex.org",
                      verbose = FALSE,
                      ...) {
-  if (!(is.null(search) || is.null(options$sample))) {
-    stop("You can't use `search` and `sample` at the same time. Please specify only one of these two arguments.")
-  }
 
   entity <- match.arg(entity, oa_entities())
   filter <- c(filter, list(...))
@@ -556,7 +578,7 @@ oa_query <- function(filter = NULL,
     )
   } else {
     path <- paste(entity, identifier, sep = "/")
-    query <- NULL
+    query <- options
   }
 
   query_url <- httr::modify_url(